* [jsfm] split vdom.js into different files
