+++ html5/runtime/vdom/comment.js
@@ -1,0 +1,19 @@
+import Node from './node'
+import { uniqueId } from './operation'
+
+export default function Comment (value) {
+  this.nodeType = 8
+  this.nodeId = uniqueId()
+  this.ref = this.nodeId
+  this.type = 'comment'
+  this.value = value
+  this.children = []
+  this.pureChildren = []
+}
+
+Comment.prototype = new Node()
+Comment.prototype.constructor = Comment
+
+Comment.prototype.toString = function () {
+  return '<!-- ' + this.value + ' -->'
+}
