+++ html5/runtime/vdom/index.js
@@ -1,0 +1,11 @@
+import Node from './node'
+import Comment from './comment'
+import Element from './element'
+import Document from './document'
+
+export {
+  Document,
+  Node,
+  Element,
+  Comment
+}
